optimize dynamic theme
