+++ src/App.tsx
@@ -4,15 +4,13 @@ import "./styles/themes.scss";
 import { Provider } from "react-redux";
 import { store } from "./store";
 import { router } from "./pages";
-import { Theme } from "./components/Theme/Theme";
+import { useInitTheme } from "./hooks/themes/useInitTheme";
 
 function App() {
-  document.documentElement.setAttribute("data-theme", "default");
+  useInitTheme();
   return (
     <Provider store={store}>
-      <Theme>
-        <RouterProvider router={router} />
-      </Theme>
+      <RouterProvider router={router} />
     </Provider>
   );
 }
