--- src/components/Theme/ThemeContext.tsx
@@ -1,4 +1,0 @@
-import { createContext } from "react";
-import { Themes } from "../../constants/themes";
-
-export const ThemeContext = createContext<((theme: Themes) => void )| null>(null);
