+++ src/hooks/themes/useInitTheme.ts
@@ -1,0 +1,11 @@
+import { useEffect } from "react"
+import { getThemeLocalStorage, setThemeLocalStorage } from "../../utils/themeController"
+
+export const useInitTheme = () => {
+  useEffect(() => {
+    const theme = getThemeLocalStorage()
+    if (theme) {
+      setThemeLocalStorage(theme)
+    }
+  }, [])
+}
