+++ src/hooks/themes/useTheme.ts
@@ -1,0 +1,8 @@
+import {
+  getThemeLocalStorage,
+  setThemeLocalStorage,
+} from "../../utils/themeController";
+
+export const useTheme = (): [typeof getThemeLocalStorage, typeof setThemeLocalStorage] => {
+  return [getThemeLocalStorage, setThemeLocalStorage];
+};
