+++ src/utils/themeController.ts
@@ -1,18 +1,16 @@
 import { THEME_KEY, Themes } from "../constants/themes";
 
-
+export const setThemeLocalStorage = (theme: Themes) => {
+  localStorage.setItem(THEME_KEY, theme);
+  document.documentElement.setAttribute("data-theme", theme);
+  return theme;
+};
 export const getThemeLocalStorage = () => {
   let theme = localStorage.getItem(THEME_KEY);
 
   if (!theme) {
-    localStorage.setItem(THEME_KEY, Themes.Default);
-    theme = Themes.Default;
+    theme = setThemeLocalStorage(Themes.Default);
   }
 
   return theme as Themes;
-};
-
-export const setThemeLocalStorage = (theme: Themes) => {
-  localStorage.setItem(THEME_KEY, theme);
-  return theme;
 };
